backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sse-starlette
cadquery
google-generativeai
python-dotenv
Pillow
